prototype/streamlitapp/requirements.txt
streamlit==1.40.0
psycopg2-binary
pandas
streamlit-aggrid